#8 動作することを確認
