+++ frontend/src/features/admin/components/AdminPage/index.tsx
@@ -1,6 +1,3 @@
-import { useState, useEffect } from "react";
-import { useRouter } from "next/navigation";
-import { supabase } from "@/lib/supabase";
 import { Button } from "@/components/shadcn/ui/button";
 import { Input } from "@/components/shadcn/ui/input";
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/shadcn/ui/card";
@@ -11,118 +8,29 @@ interface AllowedEmail {
   created_at: string;
 }
 
-export default function AdminPageComponent() {
-  const [allowedEmails, setAllowedEmails] = useState<AllowedEmail[]>([]);
-  const [newEmail, setNewEmail] = useState("");
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [isAdmin, setIsAdmin] = useState(false);
-  const router = useRouter();
+interface AdminPageComponentProps {
+  allowedEmails: AllowedEmail[];
+  newEmail: string;
+  setNewEmail: (email: string) => void;
+  loading: boolean;
+  error: string | null;
+  isAdmin: boolean;
+  addAllowedEmail: (e: React.FormEvent) => Promise<void>;
+  removeAllowedEmail: (id: string) => Promise<void>;
+  navigateToHome: () => void;
+}
 
-  // Check if user is admin
-  useEffect(() => {
-    const checkAdmin = async () => {
-      const { data: { session } } = await supabase.auth.getSession();
-      
-      if (!session) {
-        router.push("/login");
-        return;
-      }
-
-      // Get user role from Supabase
-      const { data, error } = await supabase.rpc('is_admin');
-      
-      if (error) {
-        console.error("Error checking admin status:", error);
-        setIsAdmin(false);
-        router.push("/");
-        return;
-      }
-
-      if (!data) {
-        setIsAdmin(false);
-        router.push("/");
-        return;
-      }
-
-      setIsAdmin(true);
-      fetchAllowedEmails();
-    };
-
-    checkAdmin();
-  }, [router]);
-
-  const fetchAllowedEmails = async () => {
-    try {
-      setLoading(true);
-      const { data, error } = await supabase
-        .from('allowed_emails')
-        .select('*')
-        .order('created_at', { ascending: false });
-
-      if (error) throw error;
-      setAllowedEmails(data || []);
-    } catch (error: any) {
-      console.error("Error fetching allowed emails:", error);
-      setError(error.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const addAllowedEmail = async (e: React.FormEvent) => {
-    e.preventDefault();
-    
-    if (!newEmail.trim()) return;
-    
-    try {
-      setLoading(true);
-      setError(null);
-      
-      // Validate email format
-      const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-      if (!emailRegex.test(newEmail)) {
-        throw new Error("有効なメールアドレスを入力してください");
-      }
-
-      const { data, error } = await supabase
-        .from('allowed_emails')
-        .insert([{ email: newEmail.trim() }])
-        .select();
-
-      if (error) throw error;
-      
-      setNewEmail("");
-      fetchAllowedEmails();
-    } catch (error: any) {
-      console.error("Error adding allowed email:", error);
-      setError(error.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const removeAllowedEmail = async (id: string) => {
-    try {
-      setLoading(true);
-      setError(null);
-      
-      const { error } = await supabase
-        .from('allowed_emails')
-        .delete()
-        .eq('id', id);
-
-      if (error) throw error;
-      
-      fetchAllowedEmails();
-    } catch (error: any) {
-      console.error("Error removing allowed email:", error);
-      setError(error.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
+export default function AdminPageComponent({
+  allowedEmails,
+  newEmail,
+  setNewEmail,
+  loading,
+  error,
+  isAdmin,
+  addAllowedEmail,
+  removeAllowedEmail,
+  navigateToHome
+}: AdminPageComponentProps) {
   if (!isAdmin) {
     return null; // Will redirect in useEffect
   }
@@ -204,7 +112,7 @@ export default function AdminPageComponent() {
         <CardFooter>
           <Button
             variant="outline"
-            onClick={() => router.push("/")}
+            onClick={navigateToHome}
             className="ml-auto"
           >
             ホームに戻る
