+++ frontend/src/features/admin/containers/AdminContainer/index.tsx
@@ -1,0 +1,8 @@
+import AdminPageComponent from "@/features/admin/components/AdminPage";
+import { useAdminContainer } from "./useAdminContainer";
+
+export default function AdminContainer() {
+  const adminProps = useAdminContainer();
+  
+  return <AdminPageComponent {...adminProps} />;
+}
