+++ frontend/src/features/auth/containers/LoginContainer/index.tsx
@@ -1,0 +1,8 @@
+import LoginComponent from "@/features/auth/components/Login";
+import { useLoginContainer } from "./useLoginContainer";
+
+export default function LoginContainer() {
+  const loginProps = useLoginContainer();
+  
+  return <LoginComponent {...loginProps} />;
+}
